requirements.txt
Flask==2.3.4
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
WTForms==3.0.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
